pyproject.toml
[tool.poetry]
name = "slidemaker"
version = "0.1.0"
description = ""
authors = []
packages = [{include = "slidemaker", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
python-pptx = ">=1"

[tool.poetry.scripts]
slidemaker = "slidemaker.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "makim (>=1.29.0,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "vulture (>=2.15,<3.0)",
    "ruff (>=0.15.5,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "douki (>=0.11.0,<0.12.0)"
]
